mas-production-system/services/core/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1
openai==1.3.0
httpx==0.25.2
requests==2.31.0
prometheus_client==0.19.0
numpy==1.26.2
scipy==1.11.4
pandas==2.1.4
matplotlib==3.8.2
sympy==1.12
mpmath==1.3.0
statsmodels==0.14.0
pulp==2.7.0
networkx==3.2.1
duckduckgo_search==3.9.6
beautifulsoup4==4.12.2
faiss-cpu==1.7.4
hnswlib==0.8.0
grpcio==1.60.0
grpcio-tools==1.60.0
redis[hiredis]==5.0.1
aioredis==2.0.1
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sentry-sdk[fastapi]==1.38.0
aiolimiter==1.1.0
gunicorn==21.2.0
uvloop==0.19.0
tenacity==8.2.3

# Development and testing
bandit==1.7.5
safety==2.3.5
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
locust==2.17.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Optional scientific packages (comment out if not needed)
# astropy==6.0.0
# qutip==4.7.3
# control==0.9.4
# biopython==1.82
# pubchempy==1.0.4
# dendropy==4.6.1
# rdkit==2023.9.3
# pyscf==2.4.0
# pygame==2.5.2
# chess==1.10.0
# mido==1.3.0
# midiutil==1.2.1
# torch==2.1.1